Separate points logic into multiple modules
